+++ spec/filters/thread_filter_spec.rb
@@ -11,7 +11,7 @@ RSpec.describe Airbrake::Filters::ThreadFilter do
     subject.call(notice)
     th.thread_variable_set(:bingo, nil)
 
-    expect(notice[:params][:thread][:thread_variables][:bingo]).to eq(':bango')
+    expect(notice[:params][:thread][:thread_variables][:bingo]).to eq(:bango)
   end
 
   it "appends fiber variables" do
@@ -19,7 +19,7 @@ RSpec.describe Airbrake::Filters::ThreadFilter do
     subject.call(notice)
     th[:bingo] = nil
 
-    expect(notice[:params][:thread][:fiber_variables][:bingo]).to eq(':bango')
+    expect(notice[:params][:thread][:fiber_variables][:bingo]).to eq(:bango)
   end
 
   it "appends name", skip: !Thread.current.respond_to?(:name) do
@@ -48,5 +48,33 @@ RSpec.describe Airbrake::Filters::ThreadFilter do
   it "appends safe_level", skip: Airbrake::JRUBY do
     subject.call(notice)
     expect(notice[:params][:thread][:safe_level]).to eq(0)
+  end
+
+  context "when an IO-like object is stored" do
+    let(:io_obj) do
+      Class.new(IO) do
+        def initialize; end
+      end.new
+    end
+
+    before do
+      expect(io_obj).to be_is_a(IO)
+    end
+
+    it "doesn't append the IO object to thread variables" do
+      th.thread_variable_set(:io, io_obj)
+      subject.call(notice)
+      th.thread_variable_set(:io, nil)
+
+      expect(notice[:params][:thread][:thread_variables][:io]).to be_nil
+    end
+
+    it "doesn't append the IO object to thread variables" do
+      th[:io] = io_obj
+      subject.call(notice)
+      th[:io] = nil
+
+      expect(notice[:params][:thread][:fiber_variables][:io]).to be_nil
+    end
   end
 end